requirements.txt
gradio
whisper
translate
python-dotenv
elevenlabs